+++ R/binning.R
@@ -14,7 +14,8 @@
 #'
 #' @param size `numeric(1)` with the size of a bin.
 #'
-#' @param breaks `numeric` defining the breaks (bins).
+#' @param breaks `numeric` defining the breaks (bins). See [breaks_ppm()] to
+#'     define breaks with increasing size (depending on ppm).
 #'
 #' @param FUN `function` to be used to aggregate values of `x` falling into the
 #'     bins defined by `breaks`. `FUN` is expected to return a `numeric(1)`.
@@ -24,7 +25,7 @@
 #'     `returnMids = FALSE` might be useful if the breaks are defined before
 #'     hand and binning needs to be performed on a large set of values (i.e.
 #'     within a loop for multiple pairs of `x` and `y` values).
-#'     
+#'
 #' @param .check `logical(1)` whether to check that `y` is an ordered vector.
 #'     Setting `.check = FALSE` will improve performance, provided you are sure
 #'     that `y` is always ordered.
@@ -65,7 +66,7 @@ bin <- function(x, y, size = 1,
                 .check = TRUE) {
     if (length(x) != length(y))
         stop("lengths of 'x' and 'y' have to match.")
-    if (.check) 
+    if (.check)
         if (is.unsorted(y)) stop("'y' must be an ordered vector")
     FUN <- match.fun(FUN)
     breaks <- .fix_breaks(breaks, range(y))
@@ -97,4 +98,64 @@ bin <- function(x, y, size = 1,
         brks <- c(brks, max((rng[2] + 1e-6),
                             brks[length(brks)] + mean(diff(brks))))
     brks
+}
+
+#' @title Sequence with increasing difference between elements
+#'
+#' @description
+#'
+#' `breaks_ppm` creates a sequence of numbers with increasing differences
+#' between them. Parameter `ppm` defines the amount by which the difference
+#' between values increases. The value for an element `i+1` is calculated by
+#' adding `size` to the value of element `i` and in addition also the
+#' `ppm(a, ppm)`, where `a` is the value of the element `i` plus `size`. This
+#' iterative calculation is stopped once the value of an element is larger
+#' than `to`. The last value in the result vector will thus not be equal to
+#' `to` (which is in contrast to the base [`seq()`] function) but slightly
+#' higher.
+#'
+#' A typical use case of this function would be to calculate breaks for the
+#' binning of m/z values of mass spectra. This function allows to create
+#' m/z-relative bin sizes which better represents measurement errors observed
+#' on certain mass spectrometry instruments.
+#'
+#' @param from `numeric(1)` with the value from which the sequence should start.
+#'
+#' @param to `numeric(1)` defining the upper bound of the sequence. Note that
+#'     the last value of the result will not be equal to `to` but equal to the
+#'     first number in the sequence which is larger than this value.
+#'
+#' @param by `numeric(1)` defining the constant part of the difference by which
+#'     numbers should increase.
+#'
+#' @param ppm `numeric(1)` defining the variable part of the difference by
+#'     which numbers should increase (expressed in parts-per-million of the
+#'     values).
+#'
+#' @return `numeric` with the sequence of values with increasing differences.
+#'     The returned values include `from` and `to`.
+#'
+#' @export
+#'
+#' @author Johannes Rainer
+#'
+#' @examples
+#'
+#' res <- breaks_ppm(20, 50, by = 1, ppm = 50)
+#' res
+#'
+#' ## difference between the values increases (by ppm)
+#' diff(res)
+breaks_ppm <- function(from = 1, to = 1, by = 1, ppm = 0) {
+    l <- ceiling((to - from + 1) / by)
+    res <- rep(NA_real_, l)
+    res[1L] <- a <- from
+    i <- 2L
+    p <- 1 + ppm(1, ppm)
+    while(a < to) {
+        a <- (a + by) * p
+        res[i] <- a
+        i <- i + 1L
+    }
+    res[!is.na(res)]
 }